Restructure direcotories for new waveform models
